added some (int) casts to silence warnings on Windows

+++ spu/binaryswap/binaryswapspu.c
@@ -805,10 +805,10 @@ AccumulateScreenBBox(const GLfloat *bbox)
 	WindowInfo *window = context->currentWindow;
 	GLfloat z2 = bbox[6];
 
-	winBox.x1 = bbox[0];
-	winBox.y1 = bbox[1];
-	winBox.x2 = bbox[4];
-	winBox.y2 = bbox[5];
+	winBox.x1 = (int) bbox[0];
+	winBox.y1 = (int) bbox[1];
+	winBox.x2 = (int) bbox[4];
+	winBox.y2 = (int) bbox[5];
 
 	/* adjust depth for alpha composite */
 	binaryswap_spu.depth = z2;
